third_party/WebKit/LayoutTests/http/tests/inspector/search/source-frame-replace-expected.txt

Tests different types of search-and-replace in SourceFrame

Running replace test for /REPLACEME1/REPLACED/  (caseSensitive):
  - // REPLACE#ME1#D#
Running replace test for /REPLACEME2/REPLACED/  (caseSensitive, replaceAll):
  - // REPLACE#ME2 REPLACEME2#D REPLACED#
Running replace test for /replaceMe1/replaced/ :
  - // replace#Me1#d#
Running replace test for /replaceMe2/replaced/  (replaceAll):
  - // replace#Me2 replaceMe2#d replaced#
Running replace test for /(REPLACE)ME[38]/$1D/  (regex, caseSensitive):
  - // REPLACE#ME3#D#
Running replace test for /(REPLACE)ME[45]/$1D/  (regex, caseSensitive, replaceAll):
  - // REPLACE#ME4 REPLACEME4#D REPLACED#
Running replace test for /(replac)eMe[38]/$1d/  (regex):
  - // replac#eMe3#d#
Running replace test for /(replace)Me[45]/$1d/  (regex, replaceAll):
  - // replace#Me4 replaceMe4#d replaced#
